Various updates as a result of testing the PC client API

+++ sip/execution_control/configuration_db/redis/db_client/processing_controller_client.py
@@ -4,6 +4,7 @@ import ast
 import json
 import logging
 import os
+import datetime
 
 from jsonschema import ValidationError, validate
 
@@ -27,7 +28,16 @@ class ProcessingControllerDbClient:
     ###########################################################################
 
     def add_scheduling_block(self, config_dict):
-        """Add Scheduling Block to the database."""
+        """Add Scheduling Block to the database.
+
+        Args:
+            config_dict (dict): SBI configuration dictionary.
+
+        Raises:
+            jsonschema.ValidationError, if the SBI configuration is invalid.
+            Exception, if ??
+
+        """
         # Get schema for validation
         schema = self._get_schema()
         LOG.debug('Adding SBI with config: %s', config_dict)
@@ -66,16 +76,45 @@ class ProcessingControllerDbClient:
                                     value["status"],
                                     value["id"])
         except ValidationError:
-            print("Schema Validation Error")
             raise
 
     # #########################################################################
     # Get functions
     # #########################################################################
 
+    def get_sbi_id(self, date=None, project='sip'):
+        """Return a Scheduling Block Instance ID string.
+
+        Args:
+            date (str or datetime, optional): Date of the Scheduling Block
+                                              Instance, if not specified use
+                                              current date.
+            project (str, optional): Project
+
+        Returns:
+            str, A valid Scheduling Block Instance ID.
+
+        """
+        if date is None:
+            date = datetime.datetime.utcnow()
+
+        if isinstance(date, str):
+            _date = date
+        elif isinstance(date, datetime.datetime):
+            _date = date.strftime('%Y%m%d')
+        else:
+            raise RuntimeError('Unknown date type.')
+
+        sbi_index = self.get_num_scheduling_blocks()
+        return '{}-{}-sbi{:03d}'.format(_date, project, sbi_index)
+
     def get_scheduling_block_ids(self):
-        """Get list of scheduling block ids."""
-        # Initialise empty list
+        """Get the sorted list of Scheduling Block Instance ids.
+
+        Returns:
+            list, All Scheduling Block Instance IDs in the database
+
+        """
         scheduling_block_ids = []
 
         # Pattern used to search scheduling block ids
@@ -86,15 +125,24 @@ class ProcessingControllerDbClient:
             if 'processing_block' not in block_id:
                 id_split = block_id.split(':')[-1]
                 scheduling_block_ids.append(id_split)
-        return scheduling_block_ids
+        return sorted(scheduling_block_ids)
 
-    def get_num_scheduling_block_ids(self):
-        """Get number of scheduling blocks ids."""
+    def get_num_scheduling_blocks(self):
+        """Get the number of Scheduling Blocks Instance ids in the database.
+
+        Returns:
+            int, The number of Scheduling Block Instance IDs
+
+        """
         return len(self.get_scheduling_block_ids())
 
     def get_processing_block_ids(self):
-        """Get list of processing block ids using the processing block id."""
-        # Initialise empty list
+        """Get the sorted list of Processing Block ids.
+
+        Returns:
+            list, All Processing Block IDs in the database
+
+        """
         _processing_block_ids = []
 
         # Pattern used to search processing block ids
@@ -104,9 +152,9 @@ class ProcessingControllerDbClient:
         for block_id in block_ids:
             id_split = block_id.split(':')[-1]
             _processing_block_ids.append(id_split)
-        return _processing_block_ids
+        return sorted(_processing_block_ids)
 
-    def get_num_processing_block_ids(self):
+    def get_num_processing_blocks(self):
         """Get number of processing block ids."""
         return len(self.get_processing_block_ids())
 
@@ -121,6 +169,7 @@ class ProcessingControllerDbClient:
         block_details = self.get_block_details(_scheduling_block_ids)
         for details in block_details:
             _sub_array_ids.append(details['sub_array_id'])
+        _sub_array_ids = sorted(list(set(_sub_array_ids)))
         return _sub_array_ids
 
     def get_sub_array_sbi_ids(self, sub_array_id):
@@ -131,30 +180,51 @@ class ProcessingControllerDbClient:
                 _ids.append(details['id'])
         return _ids
 
-    def get_block_details(self, block_id, sort=False):
-        """Get details of scheduling or processing block.
+    def get_block_details(self, block_ids):
+        """Get the details of a Scheduling or Processing block.
 
         Args:
-            block_id (list): List of block IDs
-            sort (bool): Set to True to sort the list
-        """
-        # Check for any duplicates
-        block_ids = set([x for x in block_id if block_id.count(x) > 0])
+            block_ids (list): List of block IDs
 
-        if sort:
-            block_ids = sorted(block_ids)
-            print("In here")
+        Returns:
+            generator?!
+
+        """
+        # Convert input to list, if needed
+        if not isinstance(block_ids, (list, tuple)):
+            block_ids = [block_ids]
 
         for _id in block_ids:
-            block_name = self._db.get_block(_id)
-            for name in block_name:
-                blocks = self._db.get_hash_dict(name)
-                yield blocks
+            block_key = self._db.get_block(_id)[0]
+            block_data = self._db.get_hash_dict(block_key)
+            # NOTE(BM) unfortunately the following hack doesn't quite work \
+            # for keys where the value is a Python object (list, dict etc.) \
+            # but is good enough for now where objects are stored as strings.
+            for key in block_data:
+                for char in ['[', '{']:
+                    if char in block_data[key]:
+                        block_data[key] = ast.literal_eval(
+                            str(block_data[key]))
+            yield block_data
 
-    def get_latest_event(self, event_block):
-        """Get the latest event added."""
-        block_event = event_block + '_events'
-        block_history = event_block + '_event_history'
+    def get_latest_event(self, event_class):
+        """Get the latest event added.
+
+        Args:
+            event_class (str): Class of event, either 'processing_block' or
+                               'scheduling_block'
+
+        Returns:
+            dict, event data.
+
+        Raises:
+            Keyerror, if an invalid event class is specified.
+
+        """
+        if event_class not in ('processing_block', 'scheduling_block'):
+            raise KeyError('Invalid event class')
+        block_event = event_class + '_events'
+        block_history = event_class + '_event_history'
         event = self._db.get_event(block_event, block_history)
         return ast.literal_eval(event)
 
@@ -176,9 +246,22 @@ class ProcessingControllerDbClient:
         """Delete the specified Scheduling Block Instance.
 
         Removes the Scheduling Block Instance, and all Processing Blocks
-        that belong to it from the database
+        that belong to it from the database.
+
+        Args:
+            block_id (str): Scheduling Block ID
+
+        Raises:
+            KeyError: If the specified block_id is not found.
+
         """
-        scheduling_blocks = self._db.get_keys(block_id)
+        LOG.debug('Deleting SBI %s', block_id)
+        scheduling_blocks = self._db.get_all_blocks(block_id)
+
+        if not scheduling_blocks:
+            raise KeyError('Scheduling Block Instance not found: {}'
+                           .format(block_id))
+
         if scheduling_blocks:
             for blocks in scheduling_blocks:
                 if "processing_block" not in blocks:
@@ -200,11 +283,21 @@ class ProcessingControllerDbClient:
     def delete_processing_block(self, processing_block_id):
         """Delete Processing Block(s).
 
-        Uses Processing Block IDs
+        Args:
+            processing_block_id (str): Processing Block ID
+
+        Raises:
+            KeyError if the specified processing_block_id is not found.
+
         """
+        LOG.debug('Deleting Processing Block %s', processing_block_id)
         processing_block = self._db.get_block(processing_block_id)
+
+        if not processing_block:
+            raise KeyError('Processing Block not found: {}'
+                           .format(processing_block_id))
+
         for block in processing_block:
-            print("")
             if 'processing_block' in block:
                 self._db.delete_key(block)
 
@@ -213,8 +306,7 @@ class ProcessingControllerDbClient:
                 scheduling_block_details = self.get_block_details(
                     [scheduling_block_id[1]])
                 for block_details in scheduling_block_details:
-                    block_list = ast.literal_eval(
-                        block_details['processing_blocks'])
+                    block_list = block_details['processing_blocks']
                     if processing_block_id in block_list:
                         block_list.remove(processing_block_id)
                     self.update_value(scheduling_block_id[1],
@@ -223,7 +315,7 @@ class ProcessingControllerDbClient:
                 # Add a event to the processing block event list to notify
                 # about deleting from the db
                 self._db.push_event(self.processing_event_name, "deleted",
-                                    id)
+                                    processing_block_id)
 
     # #########################################################################
     # Utility functions
@@ -258,7 +350,18 @@ class ProcessingControllerDbClient:
         return scheduling_block
 
     def _split_scheduling_block(self, scheduling_block):
-        """Split the scheduling block data into multiple names."""
+        """Split the scheduling block data into multiple names.
+
+        Args:
+            scheduling_block: ??
+
+        Returns:
+            ??
+
+        Raises:
+            Exception: ???
+
+        """
         # Initialise empty list
         _scheduling_block_data = {}
         _processing_block_data = {}
@@ -270,7 +373,7 @@ class ProcessingControllerDbClient:
             if block != 'processing_blocks':
                 _scheduling_block_data[block] = values
             else:
-                # Check if there is a processing block that already exits in
+                # Check if there is a processing block that already exists in
                 # the database
                 processing_block_id = self.get_processing_block_ids()
 
@@ -278,8 +381,8 @@ class ProcessingControllerDbClient:
                     if value['id'] not in processing_block_id:
                         _processing_block_data = values
                     else:
-                        raise Exception("Processing block already exits",
-                                        value['id'])
+                        raise Exception("Processing block already exists: {}"
+                                        .format(value['id']))
 
         # Adding processing block id to the scheduling block list
         for block_id in _processing_block_data:
